Génération JWT avec vérif email/password

+++ src/main/java/com/bcefit/infrastructure/IUserLoginDataRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface IUserLoginDataRepository extends CrudRepository<UserLoginData, Integer> {
+public interface IUserLoginDataRepository extends CrudRepository<UserLoginData, Long> {
 
     UserLoginData findByEmail(String email);
 }
